+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - General:
   - Native file dialogs (awt)
   - Project panel now opens with a click as default, you can roll back to hovering in settings.
+- Grids:
+  - Grid view can be forced if header detection result is uncertain
 - External scripts:
   - Support for search output sections (specify a `sectionRegex` in *.jloga.json* file)